requirements.txt
fastapi~=0.115.14
uvicorn
pydantic~=2.11.7
python-jose[cryptography]~=3.5.0
passlib[bcrypt]~=1.7.4
SQLAlchemy~=2.0.41
alembic~=1.16.2
psycopg2~=2.9.10
db~=0.1.1
pytest~=8.4.1
httpx~=0.28.1
sqlmodel~=0.0.24